fix: create title with conditions in a separated component
