+++ frontend/components/CardBoard/CardBody/CardTitle/partials/Title/index.tsx
@@ -1,0 +1,33 @@
+import Link from 'next/link';
+
+import { StyledBoardTitle } from './styles';
+
+type Props = {
+	userIsParticipating: boolean;
+	boardId: string;
+	title: string;
+	isSubBoard: boolean | undefined;
+	mainBoardId?: string;
+};
+
+const Title = ({ userIsParticipating, boardId, title, isSubBoard, mainBoardId }: Props) => {
+	return userIsParticipating ? (
+		<Link
+			key={boardId}
+			href={{
+				pathname: `boards/[boardId]`,
+				query: isSubBoard ? { boardId, mainBoardId } : { boardId }
+			}}
+		>
+			<StyledBoardTitle data-disabled={!userIsParticipating}>{title}</StyledBoardTitle>
+		</Link>
+	) : (
+		<StyledBoardTitle data-disabled={!userIsParticipating}>{title}</StyledBoardTitle>
+	);
+};
+
+Title.defaultProps = {
+	mainBoardId: undefined
+};
+
+export { Title };
