+++ frontend/components/CardBoard/CardBody/CardTitle/partials/Title/styles.tsx
@@ -1,0 +1,21 @@
+import { styled } from '../../../../../../stitches.config';
+import Text from '../../../../../Primitives/Text';
+
+const StyledBoardTitle = styled(Text, {
+	fontWeight: '$bold',
+	fontSize: '$14',
+	letterSpacing: '$0-17',
+	'&[data-disabled="true"]': { opacity: 0.4 },
+	'@hover': {
+		'&:hover': {
+			'&[data-disabled="true"]': {
+				textDecoration: 'none',
+				cursor: 'default'
+			},
+			textDecoration: 'underline',
+			cursor: 'pointer'
+		}
+	}
+});
+
+export { StyledBoardTitle };
